feat: allow running programmatically
